+++ mutwo/core_utilities/exceptions.py
@@ -98,5 +98,5 @@ class NoSolutionFoundError(Exception):
 
 
 class EmptyEnvelopeError(Exception):
-    def __init__(self, envelope):
-        super().__init__(f"Can't call 'value_at' on empty envelope '{envelope}'!")
+    def __init__(self, envelope, method):
+        super().__init__(f"Can't call '{method}' on empty envelope '{envelope}'!")